requirements.txt
Django==6.0
django-htmx>=1.27.0
psycopg[binary]>=3.1
gunicorn>=21.2
Faker==40.1.2
Pillow==12.1.0
numpy>=2.4,<2.5
python-dateutil>=2.8.2
reportlab>=4.2